@@ -19,14 +19,14 @@ class BluetoothDevice {
 
   /// Establishes a connection to the Bluetooth Device.
   Future<void> connect({
-    Duration timeout,
+    Duration? timeout,
     bool autoConnect = true,
   }) async {
     var request = protos.ConnectRequest.create()
       ..remoteId = id.toString()
       ..androidAutoConnect = autoConnect;
 
-    Timer timer;
+    Timer? timer;
     if (timeout != null) {
       timer = Timer(timeout, () {
         disconnect();
